+++ CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ```
 ## [tag] - YYYY-MM-DD
-[tag]: https://github.com/eureka-framework/component-passwprd/compare/4.2.0...master
+[tag]: https://github.com/eureka-framework/component-password/compare/5.0.0...master
 ### Changed
  - Change 1
 ### Added
@@ -16,9 +16,26 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ```
 ----
 
+## [5.0.0] - 2023-06-14
+[5.0.0]: https://github.com/eureka-framework/component-password/compare/4.2.0...5.0.0
+### Changed
+- Now compatible with PHP 8.3
+- Update Makefile
+- Update composer.json
+- Update GitHub workflow
+- Some minor code change
+- Now use eureka console 6.0
+- Move script file to dedicated root directory
+### Added
+- Add php-cs-fixer as linter
+### Removed
+- Drop PHP 7.4 & 8.0 support
+- Remove PHPCS dependency
+
+----
 
 ## [4.2.0] - 2023-06-14
-[4.2.0]: https://github.com/eureka-framework/component-passwprd/compare/4.1.0...4.2.0
+[4.2.0]: https://github.com/eureka-framework/component-password/compare/4.1.0...4.2.0
 ### Changed
 - Now compatible with PHP 8.2
 - Update Makefile
@@ -49,6 +66,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
  - Set up CI
  - Suggest eureka/component-console as complementary package for password script helper
 
+----
 
 ## [3.x.y] Release v3.x.y
 ### Changed